Rename about:style to about:styles
Auditors: @jkup

Test Plan:
- open new tab, type about:styles, verify it loads
- visit about:about + click about:styles, verify that loads
- run `npm run test -- --grep="about:styles"`
