+++ app/extensions/brave/about-styles.html
@@ -12,7 +12,7 @@
     <title data-l10n-id="styleGuide"></title>
     <script src='js/about.js'></script>
     <script src="ext/l20n.min.js" async></script>
-    <link rel="localization" href="locales/{locale}/style.properties">
+    <link rel="localization" href="locales/{locale}/styles.properties">
  </head>
  <body>
     <div id="appContainer"/>
