+++ app/extensions/brave/locales/en-US/styles.properties
